requirements.txt
flet
neurum
googlesearch-python
bs4
requests